Update
Announcement char reduced to 500. When unhiding a post/item, it will create a notification for user. Need to email the user when hiding an item/post for them to know necessary details for appealing their post. Also, besides the confirmation in the app notifications, user should be notified in their emails as well